fix: update all imports and remove dtype param across repo
All files now use `from neuron_steer.core import NeuronSteerer` (matching
package structure from PR #1) and remove the dtype= constructor arg
(bfloat16 is now hardcoded in core.py).

+++ examples/interactive_demo.py
@@ -2,8 +2,7 @@
 
 Commands: prompt, discover, ablate, amplify, sweep, edges, top, save, load, info, quit
 """
-import torch
-from neuron_steer import NeuronSteerer
+from neuron_steer.core import NeuronSteerer
 
-steerer = NeuronSteerer("meta-llama/Llama-3.1-8B-Instruct", dtype=torch.bfloat16)
+steerer = NeuronSteerer("meta-llama/Llama-3.1-8B-Instruct")
 steerer.interactive()
